+++ Changelog.rst
@@ -9,6 +9,8 @@ Under development
 * GH#814: RSA keys for PSS can be imported.
 * GH#810: fixed negation of Ed25519 points
 * Add support for Curve25519 / X25519
+* Add attribute ``curve`` to EccPoint and EccXPoint classes,
+  with the canonical name of the curve.
 
 3.20.0 (9 January 2024)
 ++++++++++++++++++++++++++